🚚 Rename: 라우팅 변경
